+++ tests/unit/test_opentelemetry_tracing.py
@@ -15,12 +15,11 @@
 import datetime
 import importlib
 import sys
-
-import mock
+from unittest import mock
 
 try:
     import opentelemetry
-except ImportError:  # pragma: NO COVER
+except ImportError:
     opentelemetry = None
 
 if opentelemetry is not None: